# colblk: implement synthetic prefix and suffix