--- a/internal/crdbtest/crdbtest.go
+++ b/internal/crdbtest/crdbtest.go
@@ -155,7 +155,9 @@ func EncodeTimestamp(key []byte, walltime uint64, logical uint32) []byte {
 }
 
 // DecodeTimestamp decodes a MVCC timestamp from a serialized MVCC key.
-func DecodeTimestamp(mvccKey []byte) ([]byte, []byte, uint64, uint32) {
+func DecodeTimestamp(
+	mvccKey []byte,
+) (prefix []byte, untypedSuffix []byte, wallTime uint64, logicalTime uint32) {
 	tsLen := int(mvccKey[len(mvccKey)-1])
 	keyPartEnd := len(mvccKey) - tsLen
 	if keyPartEnd < 0 {
@@ -450,10 +452,3 @@ func (g *cockroachKeyGen) randTimestamp() (wallTime uint64, logicalTime uint32) 
 	}
 	return wallTime, logicalTime
 }
-
-//func randCockroachKey(rng *rand.Rand, cfg KeyConfig, blockPrefix []byte) []byte {
-//	key := make([]byte, 0, cfg.PrefixLen+MaxSuffixLen)
-//	key = append(key, blockPrefix...)
-//	wallTime, logicalTime := g.rand
-//	return EncodeTimestamp(key, wallTime, cfg.Logical)
-//}