--- a/sstable/colblk_writer.go
+++ b/sstable/colblk_writer.go
@@ -959,7 +959,7 @@ func (w *RawColumnWriter) rewriteSuffixes(
 	}
 	// Copy data blocks in parallel, rewriting suffixes as we go.
 	blocks, err := rewriteDataBlocksInParallel(r, wo, l.Data, from, to, concurrency, func() blockRewriter {
-		return &colblk.DataBlockRewriter{KeySchema: wo.KeySchema}
+		return colblk.NewDataBlockRewriter(wo.KeySchema, w.comparer.Compare, w.comparer.Split)
 	})
 	if err != nil {
 		return errors.Wrap(err, "rewriting data blocks")