First commit is #3996

#### colblk: implement synthetic prefix and suffix

`DataBlockIter` now takes into account the synthetic prefix and suffix
transforms.