@RaduBerinde 2024-10-04
Merged with the recent `DataBlockIter` changes, PTAL.