@RaduBerinde 2024-10-06
*[`sstable/colblk/data_block.go` line 1026 at r5](https://reviewable.io/reviews/cockroachdb/pebble/3997#-O8Y-9wbBZuO-udj7H-W:-O8YHzaDCktBbgpjxaLx:bzid3ue) ([raw file](https://github.com/cockroachdb/pebble/blob/6744ef2902cb414e9275c5c2cec25a580aa71871/sstable/colblk/data_block.go#L1026)):*
<details><summary><i>Previously, jbowens (Jackson Owens) wrote…</i></summary><blockquote>

code coverage is saying the fast path isn't covered by our tests; is that right? I thought our datadriven unit tests covered the no transform SeekGE case?
</blockquote></details>

Wow, great catch!! I missed a few ! in NoTransforms()


<!-- Sent from Reviewable.io -->